requirements.txt
flask==3.0.3
werkzeug==3.0.3
pywebview>=5.1.0
psutil>=6.0.0
